json

+++ app/src/main/java/com/yakymovych/simon/telegramchart/Model/local/Plot.java
@@ -3,7 +3,7 @@ package com.yakymovych.simon.telegramchart.Model.local;
 import java.util.List;
 
 public class Plot {
-    public List<Integer> x,prx;
+    public List<Long> x,prx;
     public List<Double> y,pry;
     public String color = "";
 }
